requirements.txt
ultralytics
fastapi
jupyter
ipykernel
uvicorn[standard]
SQLAlchemy
pymysql
python-dotenv
passlib[bcrypt]
PyJWT
ipykernel
python-multipart